Перевести на использование Fragment() для экрана "Медиатека"
1. Осуществлён переход на Fragment() для экрана "Медиатека"
2. Убрана копка "Назад" из Toolbar, т.к. в дальнейшем для данного экрана будет использоваться BottomNavigationView
