+++ app/src/main/java/com/practicum/playlistmaker/media/presentation/adapter/MediaViewPagerAdapter.kt
@@ -1,12 +1,16 @@
 package com.practicum.playlistmaker.media.presentation.adapter
 
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.practicum.playlistmaker.media.presentation.view.FavoriteFragment
-import com.practicum.playlistmaker.media.presentation.view.MediaActivity
 import com.practicum.playlistmaker.media.presentation.view.PlaylistFragment
 
-class MediaViewPagerAdapter(fragmentActivity: MediaActivity) : FragmentStateAdapter(fragmentActivity) {
+class MediaViewPagerAdapter(
+    fragmentManager: FragmentManager,
+    lifecycle: Lifecycle,
+) : FragmentStateAdapter(fragmentManager, lifecycle) {
 
     override fun getItemCount() = TAB_COUNT
 
@@ -14,7 +18,7 @@ class MediaViewPagerAdapter(fragmentActivity: MediaActivity) : FragmentStateAdap
         return when (position) {
             0 -> FavoriteFragment.newInstance()
             1 -> PlaylistFragment.newInstance()
-            else -> throw IllegalArgumentException("Invalid position")
+            else -> throw IllegalArgumentException("Invalid position: $position")
         }
     }
