+++ app/src/main/res/layout/fragment_media.xml
@@ -5,15 +5,14 @@
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/PrimeLinearLayoutStyle"
     android:id="@+id/media"
-    tools:context=".media.presentation.view.MediaActivity">
+    tools:context=".media.presentation.view.MediaFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         style="@style/CustomToolbarStyle"
         android:id="@+id/topAppBar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
-        app:title="@string/media_library_button"
-        app:navigationIcon="@drawable/ic_back"/>
+        app:title="@string/media_library_button"/>
 
     <com.google.android.material.tabs.TabLayout
         style="@style/CustomTabLayoutStyle"
